Add BuyMeCrypto Component with its own route
